Update graph computation

+++ spectraclass/graph/cpu.py
@@ -76,7 +76,7 @@ class cpActivationFlow(ActivationFlow):
                 self.nodes = nodes_data
                 self.getGraph()
                 self.I: np.ndarray = self._knn_graph.neighbor_graph[0]
-                self.D: np.ndarray = self._knn_graph.neighbor_graph[1].astype(np.int32)
+                self.D: np.ndarray = self._knn_graph.neighbor_graph[1].astype(np.float32)
                 print(f" --->  $$$D: setNodeData D=> {self.D.__class__}:{self.D.dtype}")
                 dt = (time.time()-t0)
                 print( f"Computed NN Graph with {self._knn_graph.n_neighbors} neighbors and {nodes_data.shape[0]} verts in {dt} sec ({dt/60} min)")
